refactor(utils): delete egomimicUtils.py
Completes the breakup: the compatibility shim is gone and the module no longer
exists.

The 9 files that imported from it now import from the module that actually owns
each symbol -- pose_utils, viz_utils, tensor_utils, utils, metrics. Each import
statement was rewritten by mapping every imported name through the shim's own
re-export table, so aliases are preserved and nothing is guessed:

  algo/hpt.py                                        models/hpt_nets.py
  eval/eval_act.py                                   robot/rollout.py
  eval/eval_hpt.py                                   scripts/data_visualization.py
  scripts/aria_process/aria_to_zarr.py               scripts/eva_process/eva_to_zarr.py
  scripts/data_visualization/inspector_lib/dataset_view.py

The 24 symbols that had no home and no caller are removed with the file. Each
was checked individually for any reference in .py, .yaml or .ipynb outside
egomimicUtils itself -- all 24 came back with zero. This is the check that
previously caught get_vector_from_yaw_pitch, which had no external references
but was called by get_gaze_endpoint, and so was moved rather than dropped.

Verified: no code reference to egomimicUtils remains anywhere (only three
explanatory comments naming the old module); ruff F821/F811 clean across
egomimic/*.py; 13/14 affected modules import. The 14th, egomimic.robot.rollout,
fails on the unrelated bare robot_utils import fixed separately in #563.
