+++ egomimic/scripts/eva_process/eva_to_zarr.py
@@ -17,7 +17,7 @@ from egomimic.rldb.embodiment.eva import Eva
 from egomimic.rldb.zarr.zarr_writer import ZarrWriter
 from egomimic.scripts.eva_process.eva_utils import EvaHD5Extractor
 from egomimic.utils.aws.aws_sql import timestamp_ms_to_episode_hash
-from egomimic.utils.egomimicUtils import str2bool
+from egomimic.utils.utils import str2bool
 from egomimic.utils.pose_utils import xyzw_to_wxyz
 from egomimic.utils.video_utils import resize_video_thwc, save_preview_mp4
