--- egomimic/utils/egomimicUtils.py
@@ -1,614 +1,0 @@
-"""Legacy grab-bag. Moved symbols are re-exported for compatibility.
-
-Symbols that found a proper home now live in pose_utils / viz_utils /
-tensor_utils / utils / metrics and are re-exported below so existing
-imports keep working. What remains below has no home yet and no caller;
-removing it is handled separately.
-"""
-from egomimic.utils.metrics import (  # noqa: F401
-    frechet_gaussian_over_time,
-    reverse_kl_from_samples,
-)
-from egomimic.utils.pose_utils import (  # noqa: F401
-    base_frame_to_cam_frame,
-    cam_frame_to_base_frame,
-    cam_frame_to_cam_pixels,
-    ee_pose_to_cam_frame,
-    get_vector_from_yaw_pitch,
-    interpolate_arr,
-    interpolate_arr_euler,
-    pose_to_transform,
-    transform_to_pose,
-)
-from egomimic.utils.tensor_utils import (  # noqa: F401
-    EinOpsRearrange,
-    get_sinusoid_encoding_table,
-)
-from egomimic.utils.utils import (  # noqa: F401
-    STD_SCALE,
-    download_from_huggingface,
-    fmt,
-    str2bool,
-)
-from egomimic.utils.viz_utils import (  # noqa: F401
-    draw_actions,
-    draw_dot_on_frame,
-    get_gaze_endpoint,
-)
-
-import argparse
-import math
-import os
-from numbers import Number
-from pathlib import Path
-import cv2
-import einops
-import huggingface_hub
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-import pyarrow.parquet as pq
-import pytorch_kinematics as pk
-import scipy
-import torch
-import torch.nn as nn
-import torchvision.transforms.functional as TF
-import torchvision.transforms.v2.functional as TVTF
-from scipy.spatial.transform import Rotation
-import egomimic
-
-def draw_annotation_text(
-    image: np.ndarray,
-    annotation: str,
-    font_scale: float = 0.45,
-    color: tuple = (255, 255, 255),
-    thickness: int = 1,
-) -> np.ndarray:
-    """
-    Draws annotation text on an image.
-
-    Args:
-        image (np.ndarray): Image of shape (H, W, 3) in uint8 format.
-        annotation (str): Annotation text to draw.
-        font_scale (float): Font size.
-        color (tuple): Text color (B, G, R).
-        thickness (int): Line thickness.
-
-    Returns:
-        image (np.ndarray): Annotated image.
-    """
-    font = cv2.FONT_HERSHEY_SIMPLEX
-    text = f"Annotation: {annotation}"
-    text_size, baseline = cv2.getTextSize(text, font, font_scale, thickness)
-    x, y = 10, image.shape[0] - baseline - 10
-
-    image = np.ascontiguousarray(image.copy())
-
-    image = cv2.putText(
-        image,
-        text,
-        (x, y),
-        font,
-        font_scale,
-        color,
-        thickness,
-    )
-
-    return image
-
-def draw_rotation_text(
-    image: np.ndarray,
-    gt_rot: torch.Tensor,
-    pred_rot: torch.Tensor,
-    position: tuple = (340, 20),
-    font_scale: float = 0.45,
-    color: tuple = (255, 255, 255),
-    thickness: int = 1,
-) -> np.ndarray:
-    """
-    Draws ground truth and predicted rotation vectors on an image.
-
-    Args:
-        image (np.ndarray): Image of shape (H, W, 3) in uint8 format.
-        gt_rot (torch.Tensor): Rotation vector, shape (3,) or (6,) (dual arm).
-        pred_rot (torch.Tensor): Same shape as gt_rot.
-        position (tuple): Top-left corner (x, y) for drawing text.
-        font_scale (float): Font size.
-        color (tuple): Text color (B, G, R).
-        thickness (int): Line thickness.
-
-    Returns:
-        image (np.ndarray): Annotated image.
-    """
-    font = cv2.FONT_HERSHEY_SIMPLEX
-    x, y = position
-
-    image = np.ascontiguousarray(image.copy())
-
-    if gt_rot.shape[-1] == 3:
-        image = cv2.putText(
-            image,
-            f"GT rot:    [{fmt(gt_rot)}]",
-            (x, y),
-            font,
-            font_scale,
-            color,
-            thickness,
-        )
-        image = cv2.putText(
-            image,
-            f"Pred rot:  [{fmt(pred_rot)}]",
-            (x, y + 20),
-            font,
-            font_scale,
-            color,
-            thickness,
-        )
-    elif gt_rot.shape[-1] == 6:
-        image = cv2.putText(
-            image,
-            f"L GT rot:  [{fmt(gt_rot[0:3])}]",
-            (x, y),
-            font,
-            font_scale,
-            color,
-            thickness,
-        )
-        image = cv2.putText(
-            image,
-            f"L Pred rot:[{fmt(pred_rot[0:3])}]",
-            (x, y + 20),
-            font,
-            font_scale,
-            color,
-            thickness,
-        )
-        image = cv2.putText(
-            image,
-            f"R GT rot:  [{fmt(gt_rot[3:6])}]",
-            (x, y + 40),
-            font,
-            font_scale,
-            color,
-            thickness,
-        )
-        image = cv2.putText(
-            image,
-            f"R Pred rot:[{fmt(pred_rot[3:6])}]",
-            (x, y + 60),
-            font,
-            font_scale,
-            color,
-            thickness,
-        )
-    else:
-        raise ValueError(f"Unsupported rotation shape: {gt_rot.shape}")
-
-    return image
-
-def is_key(x):
-    return hasattr(x, "keys") and callable(x.keys)
-
-def is_listy(x):
-    return isinstance(x, list)
-
-def nds_pq(file_path):
-    """
-    Open a .parquet file and explore its structure, including nested datasets.
-    """
-    try:
-        parquet_file = pq.ParquetFile(file_path)
-        print(f"File Schema:\n{parquet_file.schema}\n")
-
-        df = pd.read_parquet(file_path)
-
-        print(f"Headers (Columns): {list(df.columns)}")
-        print(f"Shape (Rows, Columns): {df.shape}")
-
-        nested_columns = []
-        for column in df.columns:
-            # Check for nested data
-            if isinstance(df[column].iloc[0], (dict, list)):
-                nested_columns.append(column)
-
-        if nested_columns:
-            print(f"Nested Headers: {nested_columns}")
-        else:
-            print("No nested headers found.")
-    except Exception as e:
-        print(f"An error occurred: {e}")
-
-def nds(nested_ds, tab_level=0):
-    """
-    Print the structure of a nested dataset.
-    nested_ds: a series of nested dictionaries and iterables.  If a dictionary, print the key and recurse on the value.  If a list, print the length of the list and recurse on just the first index.  For other types, just print the shape.
-    """
-    # print('--' * tab_level, end='')
-    if is_key(nested_ds):
-        print("dict with keys: ", nested_ds.keys())
-    elif is_listy(nested_ds):
-        print("list of len: ", len(nested_ds))
-    elif nested_ds is None:
-        print("None")
-    elif isinstance(nested_ds, Number):
-        print("Number: ", nested_ds)
-    elif isinstance(nested_ds, np.ndarray) or isinstance(nested_ds, torch.Tensor):
-        # print('\t' * (tab_level), end='')
-        print(nested_ds.shape)
-    else:
-        print("Type: ", type(nested_ds))
-
-    if is_key(nested_ds):
-        for key, value in nested_ds.items():
-            print("\t" * (tab_level), end="")
-            print(f"{key}: ", end="")
-            nds(value, tab_level + 1)
-    elif isinstance(nested_ds, list):
-        print("\t" * tab_level, end="")
-        print("Index[0]", end="")
-        nds(nested_ds[0], tab_level + 1)
-
-def ee_orientation_to_cam_frame(ee_orientation_base, T_cam_base):
-    """
-    ee_orientation_base: (N, 3, 3) rotation matrices representing orientations in the base frame.
-    T_cam_base: (4, 4) transformation matrix from base to camera.
-    returns ee_orientation_cam: (N, 3, 3) orientations in the camera frame, and Euler Angles (yaw, pitch, roll) - (N, 3)
-    """
-    T_base_cam = np.linalg.inv(T_cam_base)  # Inverse transformation
-    # Transform orientations
-    R_cam_base = T_base_cam[:3, :3]  # Extract rotation matrix from transformation
-    ee_orientation_cam = np.array(
-        [R_cam_base @ R for R in ee_orientation_base.cpu().numpy()]
-    )  # Loop over each orientation
-    ## get yaw, pitch, roll
-    batched_ypr = batched_rotation_matrices_to_euler_angles(
-        torch.tensor(ee_orientation_cam)
-    )
-    return ee_orientation_cam, batched_ypr
-
-def batched_rotation_matrices_to_euler_angles(batch_R):
-    """
-    Convert batched rotation matrices to Euler angles (ZYX order).
-    Parameters:
-        batch_R (torch.Tensor): A batched tensor of shape [batch_size, 3, 3]
-    Returns:
-        torch.Tensor: A tensor of Euler angles of shape [batch_size, 3] (yaw, pitch, roll)
-    """
-    # Reshape the tensor to merge batch and sequence dimensions for processing
-    batch_size, _, _ = batch_R.shape
-    is_torch_tensor = isinstance(batch_R, torch.Tensor)
-    if is_torch_tensor:
-        # PyTorch reshape
-        reshaped_R = batch_R.view(-1, 3, 3).cpu().numpy()
-    else:
-        # NumPy reshape
-        reshaped_R = batch_R.reshape(-1, 3, 3)
-    # reshaped_R = batch_R.view(-1, 3, 3).cpu().numpy()
-    # Use scipy's Rotation to convert rotation matrices to Euler angles
-    rotation_objects = Rotation.from_matrix(reshaped_R)
-    euler_angles = rotation_objects.as_euler(
-        "ZYX", degrees=False
-    )  # Shape [batch_size * seq_len, 3]
-    # Convert back to torch and reshape to original batch dimensions
-    euler_angles = torch.tensor(euler_angles, device=batch_R.device)
-    euler_angles = euler_angles.view(batch_size, 3)
-    return euler_angles
-
-def pose_transform(a_pose, T_a_b):
-    """
-    a_pose: (N, 3) series of poses in frame a
-    T_a_b: (4, 4) transformation matrix from frame a to frame b
-
-    returns b_pose: (N, 3) series of poses in frame b
-    """
-    orig_shape = list(a_pose.shape)
-    a_pose = a_pose.reshape(-1, 3)
-    N, _ = a_pose.shape
-    a_pose = np.concatenate([a_pose, np.ones((N, 1))], axis=1)
-
-    ee_pose_grip_cam = T_a_b @ a_pose.T
-    orig_shape[-1] += 1
-    return ee_pose_grip_cam.T.reshape(orig_shape)
-
-def ee_pose_to_cam_pixels(ee_pose_base, T_cam_base, intrinsics):
-    """
-    ee_pose_base: (N, 3)
-    T_cam_base: (4, 4)
-    intrinsics: (3, 4)
-
-
-    returns ee_pose_cam_pixels (N, 2)
-    """
-    N, _ = ee_pose_base.shape
-    ee_pose_base = np.concatenate([ee_pose_base, np.ones((N, 1))], axis=1)
-
-    ee_pose_grip_cam = np.linalg.inv(T_cam_base) @ ee_pose_base.T
-
-    px_val = intrinsics @ ee_pose_grip_cam
-    px_val = px_val / px_val[2, :]
-
-    return px_val.T
-
-def general_norm(array, min_val, max_val, arr_min=None, arr_max=None):
-    if arr_min is None:
-        arr_min = array.min()
-    if arr_max is None:
-        arr_max = array.max()
-
-    return (max_val - min_val) * ((array - arr_min) / (arr_max - arr_min)) + min_val
-
-def general_unnorm(array, orig_min, orig_max, min_val, max_val):
-    return ((array - min_val) / (max_val - min_val)) * (orig_max - orig_min) + orig_min
-
-def miniviewer(frame, goal_frame, location="top_right"):
-    """
-    overlay goal_frame in a corner of frame
-    frame: (H, W, C) numpy array
-    goal_frame: (H, W, C) numpy array
-    location: "top_right", "top_left", "bottom_left", "bottom_right"
-
-    return frame with goal_frame in top right corner (1/4 original size)
-
-    resize using TF
-    """
-    frame = frame.copy()
-    goal_frame = goal_frame.copy()
-    if isinstance(frame, np.ndarray):
-        frame = torch.from_numpy(frame)
-    if isinstance(goal_frame, np.ndarray):
-        goal_frame = torch.from_numpy(goal_frame)
-
-    goal_frame = goal_frame.permute((2, 0, 1))
-    frame = frame.permute((2, 0, 1))
-
-    goal_frame = TF.resize(goal_frame, (frame.shape[1] // 4, frame.shape[2] // 4))
-    if location == "top_right":
-        frame[:, : goal_frame.shape[1], -goal_frame.shape[2] :] = goal_frame
-    elif location == "top_left":
-        frame[:, : goal_frame.shape[1], : goal_frame.shape[2]] = goal_frame
-    elif location == "bottom_left":
-        frame[:, -goal_frame.shape[1] :, : goal_frame.shape[2]] = goal_frame
-    elif location == "bottom_right":
-        frame[:, -goal_frame.shape[1] :, -goal_frame.shape[2] :] = goal_frame
-    # frame[:, :goal_frame.shape[1], -goal_frame.shape[2]:] = goal_frame
-    return frame.permute((1, 2, 0)).numpy()
-
-def transformation_matrix_to_pose(T):
-    R = T[:3, :3]
-    p = T[:3, 3]
-    rotation_quaternion = Rotation.from_matrix(R).as_quat()
-    pose_array = np.concatenate((p, rotation_quaternion))
-    return pose_array
-
-class AlohaFK:
-    def __init__(self, robot="arx"):
-        if robot == "aloha":
-            urdf_path = os.path.join(
-                os.path.dirname(egomimic.__file__), "resources/model_eve.urdf"
-            )
-            self.chain = pk.build_serial_chain_from_urdf(
-                open(urdf_path).read(), "vx300s/ee_gripper_link"
-            )
-        elif robot == "arx":
-            urdf_path = Path(
-                os.path.join(
-                    os.path.dirname(egomimic.__file__), "resources/model_arx.urdf"
-                )
-            )
-            xml_bytes = urdf_path.read_bytes()
-
-            self.chain = pk.build_serial_chain_from_urdf(xml_bytes, "link6")
-
-    def fk(self, qpos):
-        if isinstance(qpos, np.ndarray):
-            qpos = torch.from_numpy(qpos)
-
-        return self.chain.forward_kinematics(qpos, end_only=True).get_matrix()[:, :3, 3]
-
-def robo_to_aria_imstyle(im):
-    im = TVTF.adjust_hue(im, -0.05)
-    im = TVTF.adjust_saturation(im, 1.2)
-    im = apply_vignette(im, exponent=1)
-
-    return im
-
-def create_vignette_mask(height, width, exponent=2):
-    """
-    Create a vignette mask with the given height and width.
-    The exponent controls the strength of the vignette effect.
-    """
-    y, x = torch.meshgrid(
-        torch.linspace(-1, 1, height), torch.linspace(-1, 1, width), indexing="ij"
-    )
-    radius = torch.sqrt(x**2 + y**2) / 2
-    mask = 1 - torch.pow(radius, exponent)
-    mask = torch.clamp(mask, 0, 1)
-    return mask
-
-def apply_vignette(image_tensor, exponent=2):
-    """
-    Apply a vignette effect to a batch of image tensors.
-    """
-    N, C, H, W = image_tensor.shape
-    vignette_mask = create_vignette_mask(H, W, exponent)
-    vignette_mask = vignette_mask.unsqueeze(0).unsqueeze(
-        0
-    )  # Add batch and channel dimensions
-    vignette_mask = vignette_mask.expand(
-        N, C, H, W
-    )  # Expand to match the batch of images
-    vignette_mask = vignette_mask.to(image_tensor.device)
-    return image_tensor * vignette_mask
-
-def add_extra_train_splits(data, split_percentages):
-    """
-    data: hdf5 file in robomimic format
-    split_percentages: list of percentages for each split, e.g. [0.7, 0.1, 0.2]
-    add key "mask/train_{split_name}" which subsamples "mask/train" by split_percentages
-    """
-    N = len(data["mask/train"][:])
-    random_order = np.random.permutation(N)
-    mask = data["mask/train"][:]
-    splits = []
-    for split in split_percentages:
-        # data[f"mask/train_{split_percentages:.2f}"] = random_order[:int(N*split)]
-        sorted_order = np.sort(random_order[: int(N * split)])
-        print(sorted_order)
-        splits.append(sorted_order)
-        print(mask[sorted_order])
-        data[f"mask/train_{int(split * 100)}%"] = mask[sorted_order]
-
-    for i in range(4):
-        print(i)
-        assert set(splits[i]).issubset(set(splits[i + 1]))
-
-def interpolate_keys(obs, keys, seq_length):
-    """
-    obs: dict with values of shape (T, D)
-    keys: list of keys to interpolate
-    seq_length: int changes shape (T, D) to (seq_length, D)
-    """
-    for k in keys:
-        v = obs[k]
-        L = v.shape[0]
-        if L == seq_length:
-            continue
-
-        if k == "pad_mask":
-            # interpolate it by simply copying each index (seq_length / seq_length_to_load) times
-            obs[k] = np.repeat(v, (seq_length // L), axis=0)
-        elif k != "pad_mask":
-            interp = scipy.interpolate.interp1d(np.linspace(0, 1, L), v, axis=0)
-            try:
-                obs[k] = interp(np.linspace(0, 1, seq_length))
-            except Exception:
-                raise ValueError(
-                    f"Interpolation failed for key: {k} with shape{k.shape}"
-                )
-
-def ypr_to_matrix(ypr):
-    """Convert yaw-pitch-roll (ZYX) to rotation matrix. ypr: (..., 3) → (..., 3, 3)"""
-    yaw, pitch, roll = ypr.unbind(-1)
-
-    cy = torch.cos(yaw)
-    sy = torch.sin(yaw)
-    cp = torch.cos(pitch)
-    sp = torch.sin(pitch)
-    cr = torch.cos(roll)
-    sr = torch.sin(roll)
-
-    # Build rotation matrix (Z-Y-X / yaw-pitch-roll)
-    Rz = torch.stack(
-        [
-            torch.stack([cy, -sy, torch.zeros_like(yaw)], dim=-1),
-            torch.stack([sy, cy, torch.zeros_like(yaw)], dim=-1),
-            torch.stack(
-                [torch.zeros_like(yaw), torch.zeros_like(yaw), torch.ones_like(yaw)],
-                dim=-1,
-            ),
-        ],
-        dim=-2,
-    )
-
-    Ry = torch.stack(
-        [
-            torch.stack([cp, torch.zeros_like(pitch), sp], dim=-1),
-            torch.stack(
-                [
-                    torch.zeros_like(pitch),
-                    torch.ones_like(pitch),
-                    torch.zeros_like(pitch),
-                ],
-                dim=-1,
-            ),
-            torch.stack([-sp, torch.zeros_like(pitch), cp], dim=-1),
-        ],
-        dim=-2,
-    )
-
-    Rx = torch.stack(
-        [
-            torch.stack(
-                [torch.ones_like(roll), torch.zeros_like(roll), torch.zeros_like(roll)],
-                dim=-1,
-            ),
-            torch.stack([torch.zeros_like(roll), cr, -sr], dim=-1),
-            torch.stack([torch.zeros_like(roll), sr, cr], dim=-1),
-        ],
-        dim=-2,
-    )
-
-    return Rz @ Ry @ Rx
-
-def matrix_to_ypr(R):
-    """Convert rotation matrix to yaw-pitch-roll (ZYX). R: (..., 3, 3) → (..., 3)"""
-    # Safe conversion for all angles
-    pitch = torch.asin(-R[..., 2, 0])
-    cos_pitch = torch.cos(pitch)
-
-    yaw = torch.atan2(R[..., 1, 0] / cos_pitch, R[..., 0, 0] / cos_pitch)
-    roll = torch.atan2(R[..., 2, 1] / cos_pitch, R[..., 2, 2] / cos_pitch)
-
-    return torch.stack([yaw, pitch, roll], dim=-1)
-
-def convert_to_cam_frame(pose_base: torch.Tensor, T_cam_base: torch.Tensor):
-    """
-    pose_base: (B, T, 6) — [x, y, z, yaw, pitch, roll] in base frame
-    T_cam_base: (4, 4) — camera-to-base transform (same for all B)
-
-    Returns:
-    pose_cam: (B, T, 6) — in camera frame
-    """
-    B, T, _ = pose_base.shape
-    device = pose_base.device
-
-    # Positions
-    pos_base = pose_base[..., :3]  # (B, T, 3)
-    ypr_base = pose_base[..., 3:]  # (B, T, 3)
-
-    # Homogeneous transform
-    ones = torch.ones((B, T, 1), device=device, dtype=pose_base.dtype)
-    pos_homo = torch.cat([pos_base, ones], dim=-1)  # (B, T, 4)
-
-    T_base_cam = torch.linalg.inv(T_cam_base).to(device)  # (4, 4)
-    pos_cam_homo = torch.einsum(
-        "ij,btj->bti", T_base_cam.float(), pos_homo.float()
-    )  # (B, T, 4)
-    pos_cam = pos_cam_homo[..., :3]  # (B, T, 3)
-
-    # Orientation
-    R_base = ypr_to_matrix(ypr_base)  # (B, T, 3, 3)
-    R_cam_base = T_base_cam[:3, :3]  # (3, 3)
-
-    R_cam = torch.einsum(
-        "ij,btjk->btik", R_cam_base.float(), R_base.float()
-    )  # (B, T, 3, 3)
-    ypr_cam = matrix_to_ypr(R_cam)  # (B, T, 3)
-
-    return torch.cat([pos_cam, ypr_cam], dim=-1)  # (B, T, 6)
-
-def transform_matrix_to_pose(mat: torch.Tensor) -> torch.Tensor:
-    """
-    Convert a (B, T, 4, 4) homogeneous transform matrix to (B, T, 6) [xyz + ypr]
-
-    Args:
-        mat: (B, T, 4, 4) torch tensor — full transform matrix
-
-    Returns:
-        pose: (B, T, 6) — xyz + yaw-pitch-roll
-    """
-    assert mat.shape[-2:] == (4, 4), "Input must be (B, T, 4, 4)"
-    B, T = mat.shape[:2]
-
-    # Translation part: (B, T, 3)
-    xyz = mat[..., :3, 3]
-
-    # Rotation matrix: (B, T, 3, 3)
-    R = mat[..., :3, :3]
-
-    # Convert to yaw-pitch-roll: (B, T, 3)
-    ypr = matrix_to_ypr(R)
-
-    # Return pose: (B, T, 6)
-    return torch.cat([xyz, ypr], dim=-1)
-
